+++ libraries/weight/weight.h
@@ -2,7 +2,7 @@
 //
 //    FILE: weight.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.2.0
+// VERSION: 0.3.0
 // PURPOSE: Collection weight conversion functions
 //     URL: https://github.com/RobTillaart/weight
 
@@ -10,9 +10,13 @@
 #include "Arduino.h"
 
 
-#define WEIGHT_LIB_VERSION        (F("0.2.0"))
+#define WEIGHT_LIB_VERSION        (F("0.3.0"))
 
 
+/////////////////////////////////////////////////////////////
+//
+//  WEIGHT CONVERSION FUNCTIONS
+//
 float lbs2kilo(float lbs)     { return lbs   * 0.45359237; };
 float kilo2lbs(float kilos)   { return kilos * 2.20462262185; };
 
@@ -42,7 +46,7 @@ float US2metric(float stone, float lbs, float ounce)
 };
 
 
-//  returns lbs;
+//  returns lbs
 float metric2US(float kilo, float &stone, float &lbs, float &ounce)
 {
   float val = kilo * 2.20462262185;
@@ -59,10 +63,10 @@ float metric2US(float kilo, float &stone, float &lbs, float &ounce)
 //
 //  WEIGHT CONVERTER CLASS
 //
-class weightConvertor
+class weightConverter
 {
 public:
-  weightConvertor()                    { _w = 0; };
+  weightConverter()                    { _w = 0; };
 
   //  SETTERS
   void setKilogram(float value = 0)    { _w = value * 1000; };
@@ -83,6 +87,28 @@ public:
   void setPoint(float value = 0)       { _w = value * 0.002; };
   void setGrain(float value = 0)       { _w = value * 0.0647989; };
   void setCarat(float value = 0)       { _w = value * 0.2; };
+
+
+  //  ADDERS
+  void addKilogram(float value = 0)    { _w += value * 1000; };
+  void addGram(float value = 0)        { _w += value; };
+  void addLBS(float value = 0)         { _w += value * 453.59237; };
+  void addStone(float value = 0)       { _w += value * 6350.29318; };
+  void addOunce(float value = 0)       { _w += value * 28.349523125; };
+  void addLongTonUK(float value = 0)   { _w += value * 1016046.9088; };
+  void addShortTonUS(float value = 0)  { _w += value * 907184.74; };
+  void addQuarterUK(float value = 0)   { _w += value * 12700.58636; };
+  void addQuarterUS(float value = 0)   { _w += value * 11339.80925; };
+  void addSlug(float value = 0)        { _w += value * 14593.903; };
+  void addTroyPound(float value = 0)   { _w += value * 373.2417216; };
+  void addTroyOunce(float value = 0)   { _w += value * 31.1034768; };
+  void addRobie(float value = 0)       { _w += value * 10; };
+  void addDram(float value = 0)        { _w += value * 1.77184519531; };
+  void addDrachme(float value = 0)     { _w += value * 3.8; };
+  void addPoint(float value = 0)       { _w += value * 0.002; };
+  void addGrain(float value = 0)       { _w += value * 0.0647989; };
+  void addCarat(float value = 0)       { _w += value * 0.2; };
+
 
   //  GETTERS
   float getKilogram()   { return _w * 0.001; }
@@ -109,5 +135,5 @@ public:
 };
 
 
-// -- END OF FILE --
+//  -- END OF FILE --